requirements-dev.txt
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
aws-cdk-lib==2.166.0
constructs==10.4.2
python-dotenv==1.0.1
mypy-boto3-dynamodb==1.35.60
